Move pytest conf from integration up a level
